01python/01python/requirements.txt
numpy==2.3.0
pandas==2.3.0
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tzdata==2025.2
pyinstaller==6.14.1
selenium==4.18.1
python-dotenv==1.0.1
pyperclip==1.8.2
